feat: Enhance home screen with monthly balance and internationalization
This commit introduces a monthly balance card to the home screen, displaying total income, expenses, and the resulting balance. It also refactors the underlying logic to handle all transaction types and adds internationalization support.

### Key Changes:

-   **Home Screen UI (`home_page.dart`):**
    -   Adds a new "Month Balance" card showing total income, total expenses, and the current balance for the month.
    -   Refactors the income and expense display into a reusable `income_expense_card.dart` widget.
    -   Improves the "no expenses" message with a more user-friendly layout.
    -   Integrates `AppLocalizations` for all static text, enabling internationalization.

-   **ViewModel & State (`home_viewmodel.dart`, `home_state.dart`):**
    -   Refactors `HomeViewModel` to fetch all transactions for the month (both income and expenses) instead of just expenses.
    -   Calculates `totalIncome`, `totalExpense`, and the overall `monthBalance`.
    -   Manages the state for the new balance figures.
    -   Changes amount handling from `BigDecimal` to `int` (cents) for better precision and performance.

-   **Use Case (`get_transactions_by_month.dart`):**
    -   Introduces `GetTransactionsByMonthUseCase` to fetch all transactions for a given month, replacing the previous expense-only use case.

-   **Localization (`app_en.arb`, `app_pt.arb`):**
    -   Adds new string resources for the home screen, including titles for balance, income, expenses, and empty states.

-   **Formatting (`currency_formatter.dart`):**
    -   Adds helper functions `formatCurrencyFromCents` and `bigDecimalToCents` to manage currency conversion from integer cents.

+++ lib/feat/home/domain/expense_category_summary.dart
@@ -1,9 +1,8 @@
-import 'package:big_decimal/big_decimal.dart';
 import 'package:finances_control/feat/transaction/domain/category.dart';
 
 class ExpenseCategorySummary {
   final Category category;
-  final BigDecimal total;
+  final int total;
   final double percentage;
 
   ExpenseCategorySummary({
